+++ mainProcess/src/controllers/backstopTestRunner.ts
@@ -1,14 +1,19 @@
-import backstop = require("backstopjs");
-import puppeteer = require("puppeteer");
+import { BackstopWorkerManager } from "./backstopWorkerManager";
 
 export class BackstopTestRunner {
-  public static setWorkingDir(path: string) {
-    process.chdir(path);
+  /* set import in getter to avoid loading them at launch time */
+  private static get puppeteer() {
+    const puppeteer = require("puppeteer");
+    return puppeteer;
   }
 
-  public static initTest(path: string) {
-    process.chdir(path);
-    return backstop('init');
+  public static setWorkingDir(workingPath: string) {
+    process.chdir(workingPath);
+  }
+
+  public static initTest(workingPath: string) {
+    process.chdir(workingPath);
+    return BackstopWorkerManager.executeCommand("init");
   }
 
   public static runTest(config: any, scenarioLabel?: string): Promise<any> {
@@ -24,7 +29,7 @@ export class BackstopTestRunner {
     }
     console.log("Test filtered with ", filterRegex);
     config = this.setPuppeteerExecutablePath(config);
-    return backstop('test', {
+    return BackstopWorkerManager.executeCommand("test", {
       filter: filterRegex,
       config
     });
@@ -39,7 +44,7 @@ export class BackstopTestRunner {
         filterRegex += "_.+_" + this.cleanFilename(viewportLabel);
       }
     }
-    return backstop('approve', {
+    return BackstopWorkerManager.executeCommand('approve', {
       config,
       filter: filterRegex
     });
@@ -50,8 +55,7 @@ export class BackstopTestRunner {
       config.engineOptions = {};
     }
     if (!config.engineOptions.executablePath) {
-      config.engineOptions.executablePath = puppeteer.executablePath().replace('app.asar', 'app.asar.unpacked');
-      console.log("changed puppeteer path to ", config.engineOptions.executablePath);
+      config.engineOptions.executablePath = this.puppeteer.executablePath().replace('app.asar', 'app.asar.unpacked');
     }
     return config;
   }