+++ mainProcess/src/eventBuses/index.ts
@@ -1,0 +1,2 @@
+import "./appEventBus";
+import "./backstopEventBus";