+++ mainProcess/src/main.ts
@@ -4,7 +4,7 @@ import { autoUpdater } from "electron-updater";
 import path = require("path");
 import "v8-compile-cache";
 
-import "./eventBus";
+import "./eventBuses";
 
 try {
   autoUpdater.checkForUpdatesAndNotify();
@@ -35,6 +35,9 @@ function createWindow() {
     mainWindow = null;
   });
 }
+
+// for notification
+app.setAppUserModelId(process.execPath);
 
 app.on('ready', createWindow);