+++ src/views/StartScreenView.vue
@@ -4,21 +4,17 @@
   </div>
 </template>
 
-<style scoped>
-
-</style>
-
 <script lang="ts">
-  import { Vue, Component } from "vue-property-decorator";
-  import OpenConfigurationComponent from "../components/OpenConfigurationComponent.vue";
+import { Vue, Component } from "vue-property-decorator";
+import OpenConfigurationComponent from "../components/start/OpenConfigurationComponent.vue";
 
-  @Component({
-    name: "start-screen-view",
-    components: {
-      OpenConfigurationComponent
-    }
-  })
-  export default class StartScreenView extends Vue {
-
+@Component({
+  name: "start-screen-view",
+  components: {
+    OpenConfigurationComponent
   }
+})
+export default class StartScreenView extends Vue {
+
+}
 </script>